Move SlidePaths to the SlidePath namespace
