+++ osu.Game.Rulesets.Sentakki/Objects/SlidePath/SlidePaths.cs
@@ -4,11 +4,10 @@ using osu.Framework.Graphics;
 using osu.Game.Rulesets.Objects;
 using osu.Game.Rulesets.Objects.Types;
 using osu.Game.Rulesets.Sentakki.Extensions;
-using osu.Game.Rulesets.Sentakki.Objects.SlidePath;
 using osu.Game.Rulesets.Sentakki.UI;
 using osuTK;
 
-namespace osu.Game.Rulesets.Sentakki.Objects;
+namespace osu.Game.Rulesets.Sentakki.Objects.SlidePath;
 
 public static class SlidePaths
 {
